requirements.txt
albumentations==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
conda==4.12.0
conda-content-trust==0+unknown
conda-package-handling==1.8.1
contourpy==1.0.5
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fairscale==0.4.10
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.3
ftfy==6.1.1
huggingface-hub==0.9.1
idna==3.3
ImageHash==4.3.1
imageio==2.21.3
importlib-metadata==4.12.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonlines==3.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.12
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==21.2.4
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycocoevalcap==1.2
pycocotools==2.0.5
pycosat==0.6.3
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
regex==2022.9.13
requests==2.27.1
ruamel-yaml-conda==0.15.100
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.10.3
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.63.0
traitlets==5.4.0
transformers==4.15.0
typing_extensions==4.3.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==4.0.3
zipp==3.8.1
